requirements.txt
psycopg2~=2.9.9
flask~=2.3.3
flask-jwt-extended==4.5.2
redis==4.6.0
argon2-cffi==23.1.0
flask-limiter==3.3.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7